requirements.txt
numpy
Pillow
opencv-python
ultralytics
streamlit